+++ ignite/templates/route/NAME.tsx.ejs
@@ -1,7 +1,7 @@
 ---
 filename: <%= props.kebabCaseName %>.tsx
 ---
-import { <%= props.pascalCaseName %>Page } from "../pages"
+import { <%= props.pascalCaseName %>Page } from "@/pages"
 
 export default function <%= props.pascalCaseName %>() {
   return <<%= props.pascalCaseName %>Page />
