+++ src/app/(app)/_layout.tsx
@@ -1,0 +1,5 @@
+import { Stack } from "expo-router"
+
+export default function AppLayout() {
+  return <Stack />
+}
