+++ src/app/(auth)/login.tsx
@@ -1,0 +1,5 @@
+import { LoginPage } from "@/pages"
+
+export default function Login() {
+  return <LoginPage />
+}
