+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export { default as NotFoundPage } from "./not-found/NotFoundPage"
+export { default as WelcomePage } from "./welcome/WelcomePage"
+export * from "./login/LoginPage"
