+++ src/pages/login/LoginPage.tsx
@@ -1,0 +1,17 @@
+import { ViewStyle } from "react-native"
+
+import { observer } from "mobx-react-lite"
+
+import { Screen, Text } from "../../components"
+
+export const LoginPage = observer(function LoginPage() {
+  return (
+    <Screen style={$root} preset="scroll">
+      <Text text="login" />
+    </Screen>
+  )
+})
+
+const $root: ViewStyle = {
+  flex: 1,
+}
